+++ pkg/query/endpointset_test.go
@@ -703,10 +703,8 @@ func TestEndpointSetUpdate_AvailabilityScenarios(t *testing.T) {
 
 		lset := e.LabelSets()
 		testutil.Equals(t, 2, len(lset))
-		testutil.Equals(t, "addr", lset[0][0].Name)
-		testutil.Equals(t, addr, lset[0][0].Value)
-		testutil.Equals(t, "a", lset[1][0].Name)
-		testutil.Equals(t, "b", lset[1][0].Value)
+		testutil.Equals(t, addr, lset[0].Get("addr"))
+		testutil.Equals(t, "b", lset[1].Get("a"))
 		assertRegisteredAPIs(t, endpoints.exposedAPIs[addr], e)
 	}
 
@@ -738,10 +736,8 @@ func TestEndpointSetUpdate_AvailabilityScenarios(t *testing.T) {
 
 	lset := st.LabelSets()
 	testutil.Equals(t, 2, len(lset))
-	testutil.Equals(t, "addr", lset[0][0].Name)
-	testutil.Equals(t, addr, lset[0][0].Value)
-	testutil.Equals(t, "a", lset[1][0].Name)
-	testutil.Equals(t, "b", lset[1][0].Value)
+	testutil.Equals(t, addr, lset[0].Get("addr"))
+	testutil.Equals(t, "b", lset[1].Get("a"))
 	testutil.Equals(t, expected, endpointSet.endpointsMetric.storeNodes)
 
 	// New big batch of endpoints.